feat: Refactor type definitions and fix data import issues
Consolidate type definitions into new src/models/ folder (types, interfaces, classes) for better organization
Correct spelling of "Interial" to "Internal" in Total-Output-Data.json files across the project
Update import paths in components, contexts, and hooks to use new models structure
Fix SbcOutputDataContext to import types from models instead of local definitions
Repair fetchTotalSbcOutput hook to properly return data object and import correct types
Resolves TypeScript errors from mismatched interfaces and improves codebase maintainability by centralizing data contracts.
