+++ Source/src/components/data/Total-Output-Data.json
@@ -1,6 +1,7 @@
 {
   "Total Exterior Linear Feet": 428087,
-  "Total Interial Linear Feet ": 2048876,
+  "Total Internal Linear Feet": 2048876,
+  "Total Linear Feet": 2476963,
   "Total Miles Built": 469,
   "Total Square Footage": 15605234
 }
