+++ Source/src/components/helpers/fetchTotalSbcOutput/_index.ts
@@ -5,21 +5,25 @@
 
 import { useState, useEffect } from 'react';
 import totalOutputData from '../../data/Total-Output-Data.json';
-import { SbcOutputData } from '../../../types';
+import { SbcOutputData } from '../../../models/types';
 
 
-  const useFetchSbcOutputData = () => {
-    const [data, setData] = useState<SbcOutputData | null>(null);
-    const [loading, setLoading] = useState<boolean>(true);
-    const [error, setError] = useState<string | null>(null);
-  
-    useEffect(() => {
-      try {
-        setData(totalOutputData);
-      } catch (err) {
-        setError('Error loading data');
-      } finally {
-        setLoading(false);
-      }
-    }, []);
-  export default useFetchSbcOutputData;
+const useFetchSbcOutputData = () => {
+  const [data, setData] = useState<SbcOutputData | null>(null);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    try {
+      setData(totalOutputData);
+    } catch (err) {
+      setError('Error loading data');
+    } finally {
+      setLoading(false);
+    }
+  }, []);
+
+  return { data, loading, error };
+};
+
+export default useFetchSbcOutputData;
