+++ Source/src/interfaces/index.ts
@@ -1,0 +1,7 @@
+import { SbcOutputData } from '../models/types';
+
+export interface SbcOutputDataContextProps {
+  data: SbcOutputData | null;
+  loading: boolean;
+  error: string | null;
+}
