+++ Source/src/models/interfaces/index.ts
@@ -1,0 +1,7 @@
+import { SbcOutputData } from '../types';
+
+export interface SbcOutputDataContextProps {
+  data: SbcOutputData | null;
+  loading: boolean;
+  error: string | null;
+}
