+++ Source/src/models/types/index.ts
@@ -33,7 +33,8 @@ export interface MapInfoCard {
 
 export interface SbcOutputData {
   "Total Exterior Linear Feet": number;
-  "Total Interial Linear Feet": number;
+  "Total Internal Linear Feet": number;
+  "Total Linear Feet": number;
   "Total Miles Built": number;
   "Total Square Footage": number;
 }
